fix metadata export
